chore: Merge the packages into a single package
