--- app-api/pyproject.toml
@@ -1,101 +1,0 @@
-[tool.poetry]
-name = "app-api"
-version = "0.1.4"
-description = ""
-license = "MIT"
-authors = ["Chinedu Ezeofor <sneidue@email.com>"]
-packages = [{ include = "api" }]
-
-[tool.poetry.dependencies]
-python = "==3.10.8"
-datasets = "^2.14.4"
-ensure = "^1.0.3"
-fastapi = "^0.90.0"
-fast-token-classifier = "==0.2.8"
-fsspec = "^2023.9.2"
-httpx = "^0.25.1"
-loguru = "^0.7.0"
-numpy = "^1.22.3"
-pydantic = "<=1.10.13"
-python-dotenv = "^1.0.0"
-rich = "^13.6.0"
-scikit-learn = "^1.2.1"
-scipy = "== 1.11.3"
-typeguard = "^2.13.3"
-uvicorn = "^0.24.0"
-
-[tool.poetry.group.dev.dependencies]
-mypy = "^1.7.0"
-mypy-extensions = "^1.0.0"
-ruff = "^0.1.4"
-types-PyYAML = "^6.0.12.12"
-
-[tool.poetry.group.test.dependencies]
-pytest = "^7.2.0"
-pytest-cov = "== 4.0.0"
-
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
-
-[tool.mypy]
-mypy_path = ["api"]
-python_version = "3.10"
-show_error_context = true
-ignore_missing_imports = true
-warn_incomplete_stub = false
-check_untyped_defs = true
-disallow_any_generics = true
-disallow_untyped_defs = false
-warn_unused_configs = true
-
-[tool.ruff]
-# Set the maximum line length to 100.
-line-length = 100
-indent-width = 4
-
-# Assume Python 3.10
-target-version = "py310"
-
-# Exclude a variety of commonly ignored directories.
-exclude = [
-    ".bzr", ".direnv", ".dockerignore", ".eggs", ".env", ".git",
-    ".gitignore", ".git-rewrite",  ".hg", ".mypy_cache", ".nox",
-    ".pants.d", ".pytype", ".ruff_cache", ".svn", ".tox", ".venv",
-    "__pypackages__", "_build", "buck-out", "build", "dist", "env",
-    "node_modules", "venv",
-]
-
-[tool.ruff.lint]
-# Add the `line-too-long` rule to the enforced rule set.
-extend-select = ["E501"]
-select = [
-    "I001", "ANN201", "ANN202", "S107", "PIE796", "T201", "PYI009",
-    "PYI012", "PYI026", "RET505", "SIM115", "F401",
-    ]
-ignore = ["E741", "PT011", "PLR1722", "D419", "D101", "D100", "D107", "ANN401", "PLC0414"]
-
-# Allow unused variables when underscore-prefixed.
-dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
-
-[tool.ruff.format]
-# Like Black, use double quotes for strings.
-quote-style = "double"
-
-# Like Black, indent with spaces, rather than tabs.
-indent-style = "space"
-
-# Like Black, respect magic trailing commas.
-skip-magic-trailing-comma = false
-
-# Like Black, automatically detect the appropriate line ending.
-line-ending = "auto"
-
-# Pytest
-[tool.pytest.ini_options]
-addopts = "--strict-markers" # Ensure that the markers are defined
-markers = [
-            "integration: Run the integration tests.",
-            "differential: Run the integration tests."
-            ]
