+++ fast-token-classifier/Makefile
@@ -5,6 +5,7 @@ help:
 	@echo "Commands:"
 	@echo "\tsetup_venv:          creates a virtual environment (REQUIRED)."
 	@echo "\trun_test:            runs the tests."
+	@echo "\trun_api:             runs the app."
 	@echo "\trun_lint_n_style:    runs linting, style and type formatting."
 	@echo "\trun_checks:          runs tests and code quality (RECOMMENDEED)."
 	@echo "\tpublish_package:     used to build and publish the package."
@@ -21,7 +22,15 @@ run_test: setup_venv
 __test__:
 	@echo
 	@echo ">>>> Running tests. <<<<"
-	pytest -svv -W ignore
+	pytest -svvW ignore
+
+run_api:
+	poetry run ${MAKE} __run_api__
+
+__run_api__:
+	@echo
+	@echo ">>>> Starting the API. <<<<"
+	uvicorn api.app:app --port 8000 --host "0.0.0.0"
 
 run_lint_n_style:
 	poetry run ${MAKE} __run_lint_n_style__
@@ -29,7 +38,8 @@ run_lint_n_style:
 __run_lint_n_style__:
 	@echo
 	@echo ">>>> Running linting. <<<<"
-	ruff fast_token_classifier --fix && mypy fast_token_classifier
+	ruff fast_token_classifier --fix && mypy fast_token_classifier \
+	&& ruff api --fix && mypy api
 
 run_checks: run_test run_lint_n_style
