+++ fast-token-classifier/fast_token_classifier/config/core.py
@@ -2,7 +2,7 @@ from functools import lru_cache
 from pathlib import Path
 from typing import Any, Optional, Union
 
-import yaml
+import yaml  # type: ignore[import]
 from typeguard import typechecked
 
 # Custom Imports
